% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_rainfall_data}
\alias{check_rainfall_data}
\title{chck the format for each station. Not really important...}
\usage{
check_rainfall_data(station)
}
\arguments{
\item{station}{A dataframe for one specific station}
}
\description{
chck the format for each station. Not really important...
}
