% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_station_data_names.R
\name{clean_station_data_names}
\alias{clean_station_data_names}
\title{Match station names of data with station name of excel file}
\usage{
clean_station_data_names(station_names_from_files)
}
\description{
Match station names of data with station name of excel file
}
