% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events_one_station.R
\name{get_events_one_station}
\alias{get_events_one_station}
\title{Get the rainfall events}
\usage{
get_events_one_station(station = NULL, n_dry = NULL, min_thresh = NULL)
}
\arguments{
\item{station}{a dataframe containing the rainfall measurements for one station}

\item{n_dry}{the time with no rain that splits two consecutive rainfall events.}

\item{min_thresh}{the minimum amount of daily rainfall which is considered to be
not due to the insensitivity of the station or some other influences?}
}
\description{
For one station get all the rainfall-events that occured in the recording
}
