% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events_ra.R
\name{get_events_ra}
\alias{get_events_ra}
\title{Get events based on rolling average}
\usage{
get_events_ra(precip, n = 24, min_thresh)
}
\arguments{
\item{n}{Number of hours for the rolling average}
}
\description{
Get events based on rolling average
}
