% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rainfall_from_gauge.R
\name{get_rainfall_from_gauge}
\alias{get_rainfall_from_gauge}
\title{Wrapper-function to extract rainfall for one point}
\usage{
get_rainfall_from_gauge(
  point = NULL,
  date_of_landslide = NULL,
  days_back = 10,
  select_station_by = "dist",
  n_stations = 2
)
}
\description{
Wrapper-function to extract rainfall for one point
}
