% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations_by_distance.R
\name{get_station_by_distance}
\alias{get_station_by_distance}
\title{Select station for point in space}
\usage{
get_station_by_distance(position, station_sf)
}
\description{
For a given location select the n nearest stations
}
