% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_information.R
\name{get_station_information}
\alias{get_station_information}
\title{Get Raingauge Station information}
\usage{
get_station_information(
  info_path = "/mnt/CEPH_PROJECTS/Proslide/SubDailyStations_WISKI.xlsx"
)
}
\description{
Get the information for the raingauages from the excel file
}
