% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rainfall_n_days_before_slide.R
\name{rainfall_n_days_back}
\alias{rainfall_n_days_back}
\title{Get the rainfall for N days back of the Landslide}
\usage{
rainfall_n_days_back(landslide_date, rainfall_data, n_days_back)
}
\description{
Get the rainfall for N days back of the Landslide
}
