% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rainfall_data.R
\name{read_rainfall}
\alias{read_rainfall}
\title{Read the station rainfall data}
\usage{
read_rainfall(
  station = NULL,
  station_data_path = "/mnt/CEPH_PROJECTS/Proslide/HourlyPrecData/Checked_HourlyRR",
  only.names = FALSE,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{station}{If you want only the data for one station}

\item{station_data_path}{Path to the directory containing the \code{txt} files
for the rainfall data from the WISKI database}

\item{only.names}{If TRUE, returns a dataframe with only the names from the stations.
The two columns are the names from the file and the name matching the excel file.}
}
\value{
A list. Each element is a datafame and contains the data for that station.
the name of each list-element is the name of the station
}
\description{
This function is called for each landslide. Each landslide can have a vector of stations, as returned by a
\code{get_station_by_*}-function. The start date must be a vector of the same length as the \code{station}-vector.
Same is true for the \code{end_date}-vector.
}
\examples{
df = read_rainfall(<path_to_data>)[[1]]
head(df)
        date     hour precip m_measure
1 2002-10-29 01:00:00      0        12
2 2002-10-29 02:00:00      0        12
3 2002-10-29 03:00:00      0        12
4 2002-10-29 04:00:00      0        12
5 2002-10-29 05:00:00      0        12
6 2002-10-29 06:00:00      0        12
}
