% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_station.R
\name{select_station}
\alias{select_station}
\title{Choose the station for a given landslide-position}
\usage{
select_station(position = NULL, method = "distance", n = 3)
}
\arguments{
\item{n}{Number of Stations (Voronoi only uses 1)}
}
\description{
Choose the station for a given landslide-position
}
