% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seperate_rainfall_events_ra.R
\name{separate_events_ra}
\alias{separate_events_ra}
\title{function to seperate rainfall-events for rolling average}
\usage{
separate_events_ra(ra, min_thresh)
}
\arguments{
\item{ra}{Vector with rolling averages of rainfall data}

\item{min_thresh}{minimum value for an hour to be considered wet}
}
\description{
function to seperate rainfall-events for rolling average
}
