% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_rain.R
\name{slice_rain}
\alias{slice_rain}
\title{Extract rainfall for a station between two timestamps}
\usage{
slice_rain(
  rainfall_data = NULL,
  start_date = NULL,
  start_hour = 0,
  end_date = NULL,
  end_hour = 23
)
}
\description{
Extract rainfall for a station between two timestamps
}
