% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rainfall_data.R
\name{read_rainfall}
\alias{read_rainfall}
\title{Read the station rainfall data}
\usage{
read_rainfall(
  station_data_path = "/mnt/CEPH_PROJECTS/Proslide/HourlyPrecData/Checked_HourlyRR",
  only.names = FALSE
)
}
\arguments{
\item{station_data_path}{Path to the directory containing the \code{txt} files
for the rainfall data from the WISKI database}
}
\value{
A list. Each element is a datafame and contains the data for that station.
the name of each list-element is the name of the station
}
\description{
Read the station rainfall data
}
\examples{
df = read_rainfall(<path_to_data>)[[1]]
head(df)
        date     hour precip m_measure
1 2002-10-29 01:00:00      0        12
2 2002-10-29 02:00:00      0        12
3 2002-10-29 03:00:00      0        12
4 2002-10-29 04:00:00      0        12
5 2002-10-29 05:00:00      0        12
6 2002-10-29 06:00:00      0        12
}
